+++ src/sdk/PnP.Core.Test/cleantestenv.copyme.ps1
@@ -71,5 +71,6 @@ Remove-PnPList -Identity TestQueryListItemsUpdateConsistency -Force
 Remove-PnPList -Identity TestQueryListsConsistency -Force
 Remove-PnPList -Identity TestQueryListsDeleteConsistency -Force
 Remove-PnPList -Identity TestQueryListsUpdateConsistency -Force
+Remove-PnPList -Identity RESTListItemPaging -Force
 
 Disconnect-PnPOnline